Extracting most of DB into its own module
