+++ app/src/main/java/com/dessalines/habitmaker/ui/components/common/Dialogs.kt
@@ -43,7 +43,7 @@ import androidx.compose.ui.semantics.testTagsAsResourceId
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.lifecycle.asLiveData
 import com.dessalines.habitmaker.R
-import com.dessalines.habitmaker.db.AppSettingsViewModel
+import com.dessalines.habitmaker.db.viewmodels.AppSettingsViewModel
 import com.dessalines.habitmaker.utils.getVersionCode
 import dev.jeziellago.compose.markdowntext.MarkdownText
