+++ app/src/main/java/com/dessalines/habitmaker/ui/components/common/Other.kt
@@ -127,7 +127,7 @@ fun HabitChipsFlowRow(
     ) {
         if (habit.context?.isNotBlank() == true) {
             HabitInfoChip(
-                text = habit.context,
+                text = habit.context ?: "",
                 icon = Icons.Default.Schedule,
             )
         }
