+++ app/src/main/java/com/dessalines/habitmaker/ui/components/habit/EditHabitScreen.kt
@@ -31,13 +31,13 @@ import androidx.compose.ui.res.stringResource
 import androidx.lifecycle.asLiveData
 import androidx.navigation.NavController
 import com.dessalines.habitmaker.R
-import com.dessalines.habitmaker.db.AppSettingsViewModel
+import com.dessalines.habitmaker.db.viewmodels.AppSettingsViewModel
 import com.dessalines.habitmaker.db.EncouragementInsert
-import com.dessalines.habitmaker.db.EncouragementViewModel
+import com.dessalines.habitmaker.db.viewmodels.EncouragementViewModel
 import com.dessalines.habitmaker.db.HabitReminderInsert
-import com.dessalines.habitmaker.db.HabitReminderViewModel
+import com.dessalines.habitmaker.db.viewmodels.HabitReminderViewModel
 import com.dessalines.habitmaker.db.HabitUpdate
-import com.dessalines.habitmaker.db.HabitViewModel
+import com.dessalines.habitmaker.db.viewmodels.HabitViewModel
 import com.dessalines.habitmaker.notifications.scheduleRemindersForHabit
 import com.dessalines.habitmaker.ui.components.common.BackButton
 import com.dessalines.habitmaker.ui.components.common.ToolTip
