+++ app/src/main/java/com/dessalines/habitmaker/ui/components/habit/habitanddetails/HabitsAndDetailScreen.kt
@@ -33,23 +33,22 @@ import androidx.compose.ui.res.stringResource
 import androidx.lifecycle.asLiveData
 import androidx.navigation.NavController
 import com.dessalines.habitmaker.R
-import com.dessalines.habitmaker.db.AppSettingsViewModel
+import com.dessalines.habitmaker.db.viewmodels.AppSettingsViewModel
 import com.dessalines.habitmaker.db.Encouragement
-import com.dessalines.habitmaker.db.EncouragementViewModel
+import com.dessalines.habitmaker.db.viewmodels.EncouragementViewModel
 import com.dessalines.habitmaker.db.Habit
 import com.dessalines.habitmaker.db.HabitCheck
 import com.dessalines.habitmaker.db.HabitCheckInsert
-import com.dessalines.habitmaker.db.HabitCheckViewModel
-import com.dessalines.habitmaker.db.HabitReminderViewModel
+import com.dessalines.habitmaker.db.viewmodels.HabitCheckViewModel
+import com.dessalines.habitmaker.db.viewmodels.HabitReminderViewModel
 import com.dessalines.habitmaker.db.HabitUpdateStats
-import com.dessalines.habitmaker.db.HabitViewModel
+import com.dessalines.habitmaker.db.viewmodels.HabitViewModel
 import com.dessalines.habitmaker.db.SettingsUpdateHideCompleted
 import com.dessalines.habitmaker.notifications.deleteRemindersForHabit
 import com.dessalines.habitmaker.notifications.scheduleRemindersForHabit
 import com.dessalines.habitmaker.utils.HabitFrequency
 import com.dessalines.habitmaker.utils.SUCCESS_EMOJIS
 import com.dessalines.habitmaker.utils.SelectionVisibilityState
-import com.dessalines.habitmaker.utils.buildVirtualDates
 import com.dessalines.habitmaker.utils.calculatePoints
 import com.dessalines.habitmaker.utils.calculateScore
 import com.dessalines.habitmaker.utils.calculateStreaks
