+++ db/src/main/java/com/dessalines/habitmaker/db/Encouragement.kt
@@ -1,8 +1,6 @@
 package com.dessalines.habitmaker.db
 
 import androidx.annotation.Keep
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.room.ColumnInfo
 import androidx.room.Dao
 import androidx.room.Entity
@@ -79,30 +77,6 @@ class EncouragementRepository(
     fun deleteForHabit(habitId: Int) = encouragementDao.deleteForHabit(habitId)
 
     fun insert(encouragement: EncouragementInsert) = encouragementDao.insert(encouragement)
-}
-
-class EncouragementViewModel(
-    private val repository: EncouragementRepository,
-) : ViewModel() {
-    fun listForHabitSync(habitId: Int) = repository.listForHabitSync(habitId)
-
-    fun getRandomForHabit(habitId: Int) = repository.getRandomForHabit(habitId)
-
-    fun deleteForHabit(habitId: Int) = repository.deleteForHabit(habitId)
-
-    fun insert(encouragement: EncouragementInsert) = repository.insert(encouragement)
-}
-
-class EncouragementViewModelFactory(
-    private val repository: EncouragementRepository,
-) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(EncouragementViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return EncouragementViewModel(repository) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
 }
 
 val sampleEncouragements =
