+++ db/src/main/java/com/dessalines/habitmaker/db/HabitCheck.kt
@@ -1,8 +1,6 @@
 package com.dessalines.habitmaker.db
 
 import androidx.annotation.Keep
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.room.ColumnInfo
 import androidx.room.Dao
 import androidx.room.Entity
@@ -88,33 +86,6 @@ class HabitCheckRepository(
         habitId: Int,
         checkTime: Long,
     ) = habitCheckDao.deleteForDay(habitId, checkTime)
-}
-
-class HabitCheckViewModel(
-    private val repository: HabitCheckRepository,
-) : ViewModel() {
-    fun listForHabit(habitId: Int) = repository.listForHabit(habitId)
-
-    fun listForHabitSync(habitId: Int) = repository.listForHabitSync(habitId)
-
-    fun insert(habitCheck: HabitCheckInsert) = repository.insert(habitCheck)
-
-    fun deleteForDay(
-        habitId: Int,
-        checkTime: Long,
-    ) = repository.deleteForDay(habitId, checkTime)
-}
-
-class HabitCheckViewModelFactory(
-    private val repository: HabitCheckRepository,
-) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(HabitCheckViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return HabitCheckViewModel(repository) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
 }
 
 val sampleHabitChecks =
