+++ db/src/main/java/com/dessalines/habitmaker/db/Utils.kt
@@ -1,0 +1,8 @@
+package com.dessalines.habitmaker.db
+
+import java.time.LocalDate
+import java.time.ZoneId
+
+const val TAG = "com.habitmaker.db"
+
+fun LocalDate.toEpochMillis() = this.atStartOfDay(ZoneId.systemDefault()).toInstant().toEpochMilli()
